feat: add clinical trials tool
